requirements.txt
selenium
requests
pyautogui
webdriver-manager
pyyaml